+++ src/growbuf.h
@@ -13,7 +13,7 @@ class GrowBuf
 {
   public:
     GrowBuf() : m_str(0), m_pos(0), m_len(0) {}
-    GrowBuf(uint initialSize) : m_pos(0), m_len(initialSize) { m_str=static_cast<char*>(malloc(m_len)); }
+    GrowBuf(size_t initialSize) : m_pos(0), m_len(initialSize) { m_str=static_cast<char*>(malloc(m_len)); }
    ~GrowBuf()         { free(m_str); }
     GrowBuf(const GrowBuf &other)
     {
@@ -49,7 +49,7 @@ class GrowBuf
       m_str = std::exchange(other.m_str,static_cast<char*>(0));
       return *this;
     }
-    void reserve(uint size) { if (m_len<size) { m_len = size; m_str = static_cast<char*>(realloc(m_str,m_len)); } }
+    void reserve(size_t size) { if (m_len<size) { m_len = size; m_str = static_cast<char*>(realloc(m_str,m_len)); } }
     void clear()      { m_pos=0; }
     void addChar(char c)  { if (m_pos>=m_len) { m_len+=GROW_AMOUNT; m_str = static_cast<char*>(realloc(m_str,m_len)); }
                         m_str[m_pos++]=c;
@@ -57,7 +57,7 @@ class GrowBuf
     void addStr(const QCString &s) {
                         if (!s.isEmpty())
                         {
-                          uint l=s.length();
+                          size_t l=s.length();
                           if (m_pos+l>=m_len) { m_len+=l+GROW_AMOUNT; m_str = static_cast<char*>(realloc(m_str,m_len)); }
                           strcpy(&m_str[m_pos],s.data());
                           m_pos+=l;
@@ -66,7 +66,7 @@ class GrowBuf
     void addStr(const std::string &s) {
                         if (!s.empty())
                         {
-                          uint l=static_cast<uint>(s.length());
+                          size_t l=s.length();
                           if (m_pos+l>=m_len) { m_len+=l+GROW_AMOUNT; m_str = static_cast<char*>(realloc(m_str,m_len)); }
                           strcpy(&m_str[m_pos],s.c_str());
                           m_pos+=l;
@@ -75,16 +75,16 @@ class GrowBuf
     void addStr(const char *s) {
                         if (s)
                         {
-                          uint l=static_cast<uint>(strlen(s));
+                          size_t l=strlen(s);
                           if (m_pos+l>=m_len) { m_len+=l+GROW_AMOUNT; m_str = static_cast<char*>(realloc(m_str,m_len)); }
                           strcpy(&m_str[m_pos],s);
                           m_pos+=l;
                         }
                       }
-    void addStr(const char *s,uint n) {
+    void addStr(const char *s,size_t n) {
                         if (s)
                         {
-                          uint l=static_cast<uint>(strlen(s));
+                          size_t l=strlen(s);
                           if (n<l) l=n;
                           if (m_pos+l>=m_len) { m_len+=l+GROW_AMOUNT; m_str = static_cast<char*>(realloc(m_str,m_len)); }
                           strncpy(&m_str[m_pos],s,n);
@@ -93,14 +93,14 @@ class GrowBuf
                       }
     char *get() { return m_str; }
     const char *get() const { return m_str; }
-    uint getPos() const   { return m_pos; }
-    void setPos(uint newPos) { m_pos = newPos; }
-    char at(uint i) const { return m_str[i]; }
+    size_t getPos() const   { return m_pos; }
+    void setPos(size_t newPos) { m_pos = newPos; }
+    char at(size_t i) const { return m_str[i]; }
     bool empty() const { return m_pos==0; }
   private:
     char *m_str;
-    uint m_pos;
-    uint m_len;
+    size_t m_pos;
+    size_t m_len;
 };
 
 #endif
